Translator Directory: Unentangle from FSI.
TYPE: Feature
LINK: OGC-474
