+++ src/onegov/translator_directory/views/login.py
@@ -1,9 +1,9 @@
-from onegov.translator_directory import _
 from onegov.core.security import Public
-from onegov.fsi.views.login import FsiLoginForm
-from onegov.translator_directory import TranslatorDirectoryApp
-from onegov.user import Auth
 from onegov.org.views.auth import handle_login as handle_login_base
+from onegov.translator_directory import _
+from onegov.translator_directory import TranslatorDirectoryApp
+from onegov.translator_directory.forms.login import LoginForm
+from onegov.user import Auth
 
 
 @TranslatorDirectoryApp.form(
@@ -11,7 +11,7 @@ from onegov.org.views.auth import handle_login as handle_login_base
     name='login',
     template='login.pt',
     permission=Public,
-    form=FsiLoginForm
+    form=LoginForm
 )
 def handle_login(self, request, form):
     # custom default redirect
